+++ resources/js/processes/import/components/ImportManager.vue
@@ -1,0 +1,18 @@
+<template>
+    <router-view></router-view>
+</template>
+
+
+<script>
+
+export default {
+    router: window.ProcessMaker.Router,
+    props: [],
+    data() {
+        return {
+        }
+    },
+    methods: {
+    },
+}
+</script>
