+++ webpack.mix.js
@@ -90,6 +90,7 @@ mix.js("resources/js/app-layout.js", "public/js")
   .js("resources/js/processes/scripts/editConfig.js", "public/js/processes/scripts")
   .js("resources/js/processes/export/index.js", "public/js/processes/export")
   .js("resources/js/processes/environment-variables/index.js", "public/js/processes/environment-variables")
+  .js("resources/js/processes/import/index.js", "public/js/processes/import")
   .js("resources/js/processes/screens/index.js", "public/js/processes/screens")
   .js("resources/js/processes/screens/edit.js", "public/js/processes/screens")
   .js("resources/js/processes/signals/index.js", "public/js/processes/signals")
